remove id from div and delegate LoginForm to its own component
