+++ src/components/login/VideoContainer/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './styles.css';
 
 const VideoContainer = () => (
-  <div id="video" className="video-container">
+  <div className="video-container">
     <div className="play-container">
       <img src="/play.png" alt="play" className="play clickable" />
       <img src="/i6.png" alt="phone" className="phone" />
