+++ src/components/navigation/Hamburger/styles.css
@@ -10,14 +10,14 @@
   user-select: none;
 }
 
-.hamburger-container a {
+.hamburger-container .hamburger-item {
   text-decoration: none;
   color: var(--color-menu-item);
   font-weight: normal;
   transition: color 0.3s ease;
 }
 
-.hamburger-container a:hover {
+.hamburger-container .hamburger-item:hover {
   font-weight: bolder;
 }
